webui/requirements.txt
# Web framework: Flask is lightweight and simple for this two-page app.
flask==3.1.3

# WebSocket support for the terminal backend (xterm.js).
flask-sock==0.7.0

# Secure cookie/session management.
itsdangerous==2.2.0

# Password hashing for the web UI login.
bcrypt==4.1.3

# MQTT client for publishing/subscribing to phev2mqtt topics.
paho-mqtt==2.1.0

# Encrypt sensitive config values at rest.
cryptography==42.0.8
